move timedelta tests to test_timedelta.py
